Move pool warnings to config
